# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AREA Server API
  version: 0.1.0
  description: |-
    The AREA Server API powers the automation platform that links actions and reactions across
    partner services. Use it to onboard users, issue secure sessions, and orchestrate OAuth-based
    integrations.

    ### Highlights
    - Account lifecycle endpoints for registration, verification, and session management.
    - OAuth helpers to initiate and complete third-party connector authorisations.
    - Metadata routes that expose the current backend build, feature flags, and compatibility notes.

    ### Getting started
    1. Create a user via `POST /v1/users` and complete the verification journey.
    2. Authenticate with either password credentials or OAuth connectors to obtain an `area_session` cookie.
    3. Call authenticated endpoints such as `GET /v1/auth/me` with the session cookie attached as a standard cookie header.

    Refer to the environment selector in the sidebar to toggle between production, staging, and local URLs.
    
    ### Rate limits & reliability
    - Default rate limit: **120 requests per minute** per authenticated session.
    - Burst tolerance: short spikes up to 300 requests per minute are accepted for 30 seconds.
    - All endpoints are reachable over HTTPS only; staging and local URLs are for testing purposes.

    ### Support & SLA
    - Incident response via the on-call rotation is available 24/7 through the status portal.
...
  contact:
...
  license:
...
  x-logo:
...
servers:
...
security:
  - sessionAuth: []
paths:
...
  /v1/components:
    get:
      summary: List service components
      description: Retrieve automation components that can be used as actions or reactions
      operationId: listComponents
      tags:
        - catalog
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: kind
          schema:
            type: string
            enum: [action, reaction]
          description: Filter components by kind
        - in: query
          name: provider
          schema:
            type: string
          description: Filter components by provider slug (for example `google`)
      responses:
        '200':
          description: Component catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentListResponse'
        '401':
          description: Authentication required
        '500':
          description: Failed to load component catalog
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: area_session
      description: Session cookie issued after authentication.
  schemas:
...
    ComponentSummary:
      type: object
      description: Minimal catalog metadata required by clients to render a component reference.
      required: [id, kind, name, displayName, provider]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [action, reaction]
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional configuration metadata describing expected parameters.
        provider:
          $ref: '#/components/schemas/ServiceProviderSummary'
    ComponentListResponse:
      type: object
      required: [components]
      properties:
        components:
          type: array
          items:
            $ref: '#/components/schemas/ComponentSummary'
    ServiceProviderSummary:
      type: object
      description: Lightweight representation of a service provider exposing components.
      required: [id, name, displayName]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
...